Conflicts:
	.gitmodules
	Agent/Agent.cc
	Agent/agent.hh
	Agent/initMotionManager.cc
	Agent/processInputCommands.cc
	Agent/standUpIfFallen.cc
	Ambulator/ambulator.hh
	Ambulator/ambulator.ih
	Ambulator/step.cc
	Option/ActionOption/actionoption.hh
	Option/ActionOption/runPolicy.cc
	Option/FSMOption/fsmoption.hh
	Option/LookAround/lookaround.hh
	Option/LookAtFeet/lookatfeet.hh
	OptionTreeBuilder/AdHocOptionTreeBuilder/buildTree.cc
	StateObject/AmbulatorState/ambulatorstate.hh
	VisualCortex/integrateImage.cc
	main.cc
	robotis

+++ ImagePassHandler/BlobDetectPass/BlobDetectPass.cc
@@ -1,6 +1,6 @@
 #include "blobdetectpass.ih"
 
-BlobDetectPass::BlobDetectPass(int imageWidth, int imageHeight, std::vector<shared_ptr<PixelLabel>> const& pixelLabels)
+BlobDetectPass::BlobDetectPass(ushort imageWidth, ushort imageHeight, std::vector<shared_ptr<PixelLabel>> const& pixelLabels)
   : d_imageHeight(imageHeight),
     d_imageWidth(imageWidth),
     d_pixelLabels(pixelLabels),
